+++ Exceptions/InvalidConnectionStringException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AzureOss\Storage\File\Share\Exceptions;
+
+/**
+ * Indicates that a connection string cannot create a File Share service client.
+ */
+class InvalidConnectionStringException extends \Exception {}